+++ docs/make.jl
@@ -1,12 +1,12 @@
-using RAGTools
+using RAG
 using Documenter
 
-DocMeta.setdocmeta!(RAGTools, :DocTestSetup, :(using RAGTools); recursive = true)
+DocMeta.setdocmeta!(RAG, :DocTestSetup, :(using RAG); recursive = true)
 
 makedocs(;
-    modules = [RAGTools],
+    modules = [RAG],
     authors = "J S <49557684+svilupp@users.noreply.github.com> and contributors",
-    sitename = "RAGTools.jl",
+    sitename = "RAG.jl",
     format = Documenter.HTML(;
         canonical = "https://github.com/JuliaGenAI/RAGTools.jl",
         edit_link = "main",